requirements.txt
python-telegram-bot==20.7
openai==1.14.2
gspread
oauth2client
pytz
SpeechRecognition
pydub
